requirements.txt
flask
flask_mysqldb

##pip install -r requirements.txt